+++ CHANGELOG.md
@@ -8,6 +8,27 @@ If you like this project and find it useful, please consider giving it a star on
   <img src="bmc-button.svg" alt="Buy me a coffee" width="120">
 </a>
 
+
+## [3.0.3] - 2025-05-??
+
+### Added
+
+- [virtual] Added virtual devices configuration mode in the Matterbridge Settings: 'Disabled', 'Light', 'Outlet', 'Switch', 'Mounted_switch'. Switch is not supported by Alexa. Mounted Switch is not supported by Apple.
+- [deviceTypes] Add waterHeater device type.
+
+### Changed
+
+- [package]: Updated dependencies.
+- [export]: Removed long deprecated Matter exports from matterbridge. Use matterbridge/matter.
+- [matterbridge]: Refactored initialize() and cleanup() methods.
+- [matterbridge]: Updated -help informations.
+
+### Fixed
+
+<a href="https://www.buymeacoffee.com/luligugithub">
+  <img src="bmc-button.svg" alt="Buy me a coffee" width="80">
+</a>
+
 ## [3.0.2] - 2025-05-14
 
 ### Added